Product can now contain multiple images and more
A product can now contain of multiple images, these can be seen on the SingleProduct.vue page. There still needs to be an option to add extra images, but more to come in the future. Apart from that, I have added responsiveness to the SingleProduct page and form validation when you try to add a review.