Added more fields to the register form
The user will need to fill in more information now. The full name is now seperated into first and last name, also telephone number has been added and a checkbox if you want to receive updates via mail are all added in this update. 

Aside from that I have added the route to the homepage when you click on 'Home'.